Migrate TensorData
